+++ Backend/app/rider-platform/rider-app/Main/rider-app.cabal
@@ -226,7 +226,6 @@ library
       Domain.Types.Person.PersonDefaultEmergencyNumber
       Domain.Types.Person.PersonDisability
       Domain.Types.Person.PersonFlowStatus
-      Domain.Types.PlaceBasedServiceConfig
       Domain.Types.Quote
       Domain.Types.SavedReqLocation
       Domain.Types.SpecialZoneQuote
@@ -281,7 +280,6 @@ library
       Storage.Beam.Person.PersonDefaultEmergencyNumber
       Storage.Beam.Person.PersonDisability
       Storage.Beam.Person.PersonFlowStatus
-      Storage.Beam.PlaceBasedServiceConfig
       Storage.Beam.Quote
       Storage.Beam.SavedReqLocation
       Storage.Beam.SpecialZoneQuote
@@ -347,7 +345,7 @@ library
       Storage.Queries.Person.PersonDisability
       Storage.Queries.Person.PersonFlowStatus
       Storage.Queries.PersonStatsExtra
-      Storage.Queries.PlaceBasedServiceConfig
+      Storage.Queries.PlaceBasedServiceConfigExtra
       Storage.Queries.Quote
       Storage.Queries.RegistrationTokenExtra
       Storage.Queries.RentalDetailsExtra
@@ -421,6 +419,7 @@ library
       Domain.Types.NotificationSoundsConfig
       Domain.Types.OnSearchEvent
       Domain.Types.PersonStats
+      Domain.Types.PlaceBasedServiceConfig
       Domain.Types.Rating
       Domain.Types.RegistrationToken
       Domain.Types.RentalDetails
@@ -469,6 +468,7 @@ library
       Storage.Beam.NotificationSoundsConfig
       Storage.Beam.OnSearchEvent
       Storage.Beam.PersonStats
+      Storage.Beam.PlaceBasedServiceConfig
       Storage.Beam.Rating
       Storage.Beam.RegistrationToken
       Storage.Beam.RentalDetails
@@ -522,11 +522,13 @@ library
       Storage.Queries.OrphanInstances.Issue
       Storage.Queries.OrphanInstances.LocationMapping
       Storage.Queries.OrphanInstances.PersonStats
+      Storage.Queries.OrphanInstances.PlaceBasedServiceConfig
       Storage.Queries.OrphanInstances.RegistrationToken
       Storage.Queries.OrphanInstances.RentalDetails
       Storage.Queries.OrphanInstances.Ride
       Storage.Queries.OrphanInstances.SearchRequest
       Storage.Queries.PersonStats
+      Storage.Queries.PlaceBasedServiceConfig
       Storage.Queries.Rating
       Storage.Queries.RegistrationToken
       Storage.Queries.RentalDetails
