+++ Backend/app/rider-platform/rider-app/Main/src-read-only/Storage/Queries/PlaceBasedServiceConfig.hs
@@ -1,0 +1,19 @@
+{-# OPTIONS_GHC -Wno-dodgy-exports #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
+{-# OPTIONS_GHC -Wno-unused-imports #-}
+
+module Storage.Queries.PlaceBasedServiceConfig (module Storage.Queries.PlaceBasedServiceConfig, module ReExport) where
+
+import qualified Domain.Types.PlaceBasedServiceConfig
+import Kernel.Beam.Functions
+import Kernel.External.Encryption
+import Kernel.Prelude
+import Kernel.Types.Error
+import Kernel.Utils.Common (CacheFlow, EsqDBFlow, MonadFlow, fromMaybeM, getCurrentTime)
+import Storage.Queries.PlaceBasedServiceConfigExtra as ReExport
+
+create :: (EsqDBFlow m r, MonadFlow m, CacheFlow m r) => (Domain.Types.PlaceBasedServiceConfig.PlaceBasedServiceConfig -> m ())
+create = createWithKV
+
+createMany :: (EsqDBFlow m r, MonadFlow m, CacheFlow m r) => ([Domain.Types.PlaceBasedServiceConfig.PlaceBasedServiceConfig] -> m ())
+createMany = traverse_ create
